(chore) move settings components to subfolder
