+++ src/renderer/components/sidebar/settings/InterfaceSettings.jsx
@@ -2,7 +2,7 @@
 
 import React, { PropTypes, Component } from 'react';
 import i18n from 'i18next';
-import SettingsActions from '../../actions/SettingsActions';
+import SettingsActions from '../../../actions/SettingsActions';
 
 class InterfaceSettings extends Component {
   constructor(props) {
