+++ src/renderer/components/sidebar/settings/RemoteSettings.jsx
@@ -3,7 +3,7 @@
 import { ipcRenderer as ipc } from 'electron';
 import React, { PropTypes, Component } from 'react';
 import i18n from 'i18next';
-import SettingsActions from '../../actions/SettingsActions';
+import SettingsActions from '../../../actions/SettingsActions';
 
 const renderRemoteLink = function renderRemoteLink(remoteLink) {
   return (
