+++ src/renderer/components/sidebar/settings/SettingsTab.jsx
